PP Size Checker
The photograph should be in colour and of the size of 2 inches x 2 inches 51 mm x 51 mm. The photo‐print should be clear and with a continuous‐tone quality. It should have a full face, front view, eyes open. Photo should present full head from the top of hair to bottom of the chin.