README.md
The photograph should be in colour and of the size of 2 inch x 2 inch 51 mm x 51 mm . The photo‐print should be clear and with a continuous‐tone quality. It should have full face, front view, eyes open. Photo should present full head from top of hair to bottom of chin.